requirements.txt
pycrypto
python-magic-bin
pefile
